+++ iso20022-core/src/main/java/com/prowidesoftware/swift/model/mx/AppHdrFactory.java
@@ -105,6 +105,9 @@ public class AppHdrFactory {
 
         if (id != null) {
             h.setMsgDefIdr(id.id());
+            if (id.getBusinessService().isPresent()) {
+                h.setBizSvc(id.getBusinessService().get());
+            }
         }
 
         h.setCreDt(XMLGregorianCalendarUtils.now());
@@ -149,6 +152,9 @@ public class AppHdrFactory {
 
         if (id != null) {
             h.setMsgDefIdr(id.id());
+            if (id.getBusinessService().isPresent()) {
+                h.setBizSvc(id.getBusinessService().get());
+            }
         }
 
         h.setCreDt(XMLGregorianCalendarUtils.now());