+++ iso20022-core/src/main/java/com/prowidesoftware/swift/model/mx/MxParseUtils.java
@@ -194,7 +194,7 @@ public class MxParseUtils {
     public static Optional<MxId> identifyMessage(final String xml) {
         Optional<String> namespace = NamespaceReader.findDocumentNamespace(xml);
         if (namespace.isPresent()) {
-            return namespace.map(MxId::new);
+            return enrichBusinessService(namespace.map(MxId::new).orElse(null), xml);
         }
 
         // if the Document does not have a namespace, try to identify the message from the header
@@ -205,13 +205,28 @@ public class MxParseUtils {
         }
         if (element.isPresent()) {
             try {
-                return Optional.of(new MxId(element.get().getElementText()));
+                return enrichBusinessService(new MxId(element.get().getElementText()), xml);
             } catch (XMLStreamException e) {
                 log.finer("Error identifying message: " + e.getMessage());
             }
         }
 
         return Optional.empty();
+    }
+
+    private static Optional<MxId> enrichBusinessService(MxId mxId, final String xml) {
+        if (mxId == null) {
+            return Optional.empty();
+        }
+        Optional<XMLStreamReader> element = NamespaceReader.findElement(xml, "BizSvc");
+        if (element.isPresent()) {
+            try {
+                mxId.setBusinessService(element.get().getElementText());
+            } catch (XMLStreamException e) {
+                log.finer("Error identifying business service: " + e.getMessage());
+            }
+        }
+        return Optional.of(mxId);
     }
 
     /**